move a lot of shared code into MetaDataset/Dataset
